Re-refactor UART classes.
