+++ adafruit_ble/uart_client.py
@@ -28,10 +28,10 @@ UART-style communication by a Central as a GATT Client
 * Author(s): Dan Halbert for Adafruit Industries
 
 """
-from bleio import Characteristic, Central
-from .uart import UART
+from bleio import Central, CharacteristicBuffer
+from .uart import NUS_SERVICE_UUID, NUS_RX_CHAR_UUID, NUS_TX_CHAR_UUID
 
-class UARTClient(UART):
+class UARTClient:
     """
     Provide UART-like functionality via the Nordic NUS service.
 
@@ -58,18 +58,10 @@ class UARTClient(UART):
     """
 
     def __init__(self, *, timeout=5.0, buffer_size=64):
-        # Since we're remote we receive on tx and send on rx. The names
-        # are from the point of view of the server.
-        super().__init__(read_characteristic=Characteristic(UART.NUS_TX_CHAR_UUID),
-                         write_characteristic=Characteristic(UART.NUS_RX_CHAR_UUID),
-                         timeout=timeout, buffer_size=buffer_size)
-
+        self._buffer_size = buffer_size
+        self._timeout = timeout
+        self._read_char = self._write_char = self._read_buffer = None
         self._central = Central()
-
-    @property
-    def connected(self):
-        """True if we are connected to a peripheral."""
-        return self._central.connected
 
     def connect(self, address, timeout):
         """Try to connect to the peripheral at the given address.
@@ -77,8 +69,77 @@ class UARTClient(UART):
         :param bleio.Address address: The address of the peripheral to connect to
         :param float/int timeout: Try to connect for timeout seconds.
         """
-        self._central.connect(address, timeout, service_uuids=(UART.NUS_SERVICE_UUID,))
+        self._central.connect(address, timeout, service_uuids=(NUS_SERVICE_UUID,))
+
+        # Connect succeeded. Get the remote characteristics we need, which were
+        # found during discovery.
+
+        for characteristic in self._central.remote_services[0].characteristics:
+            # Since we're remote we receive on tx and send on rx.
+            # The names are from the point of view of the server.
+            if characteristic.uuid == NUS_RX_CHAR_UUID:
+                self._write_char = characteristic
+            elif characteristic.uuid == NUS_TX_CHAR_UUID:
+                self._read_char = characteristic
+        if not self._write_char or not self._read_char:
+            raise OSError("Remote UART missing needed characteristic")
+        self._read_buffer = CharacteristicBuffer(self._read_char,
+                                                 timeout=self._timeout,
+                                                 buffer_size=self._buffer_size)
 
     def disconnect(self):
         """Disconnect from the peripheral."""
         self._central.disconnect()
+        self._read_char = self._write_char = self._read_buffer = None
+
+    @property
+    def connected(self):
+        """True if we are connected to a peripheral."""
+        return self._central.connected
+
+    def read(self, nbytes=None):
+        """
+        Read characters. If ``nbytes`` is specified then read at most that many bytes.
+        Otherwise, read everything that arrives until the connection times out.
+        Providing the number of bytes expected is highly recommended because it will be faster.
+
+        :return: Data read
+        :rtype: bytes or None
+        """
+        return self._read_buffer.read(nbytes)
+
+    def readinto(self, buf, nbytes=None):
+        """
+        Read bytes into the ``buf``. If ``nbytes`` is specified then read at most
+        that many bytes. Otherwise, read at most ``len(buf)`` bytes.
+
+        :return: number of bytes read and stored into ``buf``
+        :rtype: int or None (on a non-blocking error)
+        """
+        return self._read_buffer.readinto(buf, nbytes)
+
+    def readline(self):
+        """
+        Read a line, ending in a newline character.
+
+        :return: the line read
+        :rtype: int or None
+        """
+        return self._read_buffer.readline()
+
+    @property
+    def in_waiting(self):
+        """The number of bytes in the input buffer, available to be read."""
+        return self._read_buffer.in_waiting
+
+    def reset_input_buffer(self):
+        """Discard any unread characters in the input buffer."""
+        self._read_buffer.reset_input_buffer()
+
+    def write(self, buf):
+        """Write a buffer of bytes."""
+        # We can only write 20 bytes at a time.
+        offset = 0
+        while offset < len(buf):
+            self._write_char.value = buf[offset:offset+20]
+            offset += 20
